refactor: 수업 관리 모달 디렉토리 위치 수정
